`autodoc2_module_summary = False` should still print non-summary API
When `autodoc2_module_summary = False`, each module's page only
includes the "Submodules" heading at the top and no other API
documentation. I assume this is unintended, and the "summary" being
referenced in this config option should only be the tables of each
item type at the top of each module.

I think this might have been an off-by-one indentation level error?
The code to render the non-summary parts was accidentally indented to
the level of the summary check `if`.

If the existing behavior is the intended behavior of this config
option, let me know and maybe we can change this to better document
what the behavior should be.

I didn't include any tests because I didn't see a pattern for handling
combinations of config options, but I did test manually. If you want
me to add something that tests this in isolation, let me know.

Thanks!